Removing execute/execute_raw from queryable
We made an assumption earlier that an execute function should return an
`Id` in a case of an insert. We can fetch the last id from certain
databases, but it will only work with an auto-incremented integer field
using MySQL or in case of SQLite we get a row_id which might not be what
we think it is.

This change removes `execute` and `execute_raw` methods from `Queryable`.
Everything they did can be done using `query` and `query_raw`. If the
database gives us the last id, we store it to the `ResultSet` and in
case of PostgreSQL, the query should contain a `RETURNING` part to fetch
the inserted data.