Flatten the recursion in ConditionTree
We used to recurse for every element in an AND or OR tree, causing a
stack overflow with quite a small trees. This change flattens the trees
a bit to prevent recursing too much.

An example:

```rust
"foo".equals("bar").and("omg".less_than(1)).or("wtf".equals("omg"))
```

The old code would generate the following condition tree:

```rust
ConditionTree::Or(
    ConditionTree::And(
        "foo".equals("bar"),
        "omg".less_than(1),
    ),
    "wtf".equals("omg"),
)
```

And the generated SQL would be something like:

```sql
(("foo" = 'bar' AND "omg" < 1) OR "wtf" = 'omg')
```

The new version does it a bit differently:

```rust
ConditionTree::Or(vec![
    ConditionTree::And("foo".equals("bar", "omg".less_than(1))),
    "wtf".equals("omg"),
])
```

Which will generate the following SQL:

```sql
(("foo" = 'bar' AND "omg" < 1) OR "wtf" = 'omg')
```

So here we recurse only when switching from `AND` to `OR` and back. Or
have a sigular `NOT` where we must cut the scope with parentheses for
correct results.