Test and fix raw_cmd() on MySQL
The previous behaviour was to return errors only for the first
statement, without giving feedback on the success or failure of
subsequent statements. This would also leave the connection in an
invalid state for the next queries.

We now consume all result sets from the server response in `raw_cmd()`
on MySQL.

This commit also includes as part of the fix an upgrade to mysql_async
0.24.