CTEs and emulation of tuple ops for MS
Adds support of Common Table Expressions (CTE). The starting point is
the `with` method of `Select`, allowing the user to add expressions to
the queries `WITH` block.

Additionally, on MSSQL queries if certain conditions apply, the query is
converted to use the CTEs:

- If comparing a tuple (a,b) against a selection of (SELECT x, y ..)
- ... and both sides have as many variables, more than one

Then a query where we do

```sql
SELECT * FROM X
WHERE (a, b) IN (SELECT x, y FROM ..)
```

converts into the following on SQL Server.

```sql
WITH cte_0 AS (SELECT x, y FROM ..)
WHERE a IN (SELECT x FROM cte_0 WHERE b = y)
```

This should work in every position of the query. The CTE is created to
the first selection when going up the tree, adding multiple CTEs if
having more than one tuple against select comparision, upping the index
of the `cte` by one.

Additionally, when comparing tuples, e.g. in a join context, we can
do in the open source database the following:

```sql
ON (a,b) = (x,y)
```

... but in MSSQL we convert it to:

```sql
ON a = x AND b = y
```

The same conversion applies to all tuple conversions in every position,
either when comparing equals `=` or not `<>`.