A workaround for `OUTPUT` if having triggers
If a table has triggers, it is not allowed to use the `OUTPUT` statement
on `INSERT` or `MERGE` (or `UPDATE`, but we don't care about that yet).

The way Microsoft's Entity Framework goes around this wontfix issue is
through a separate table variable, inserting the `OUTPUT` items to it
first, then selecting them using a merge with the actual table.

What makes this a bit curious is how the table variable can just hold
varchars, get numbers or uuids inserted to it and we can still merge
with the actual table and with the correct types and select them out.

This additionally adds a `type_family` marker to the `Column`, if we
ever want to have a closer mapping to the reality in the table variable.
It's probably much faster to at least use something that is closer to an
integer when having integer columns, and for sure varchars for varchar
columns.

By default the table variable uses `NVARCHAR(255)`, but if the type
family is set, we can change the type.