fix(sqlite): do NOT create two db connections
This removes support for the path params "db_name". Quaint was using the `ATTACH DATABASE` statement as a hack to scope the db under some custom name.
The `ATTACH DATABASE` statement is initially designed to create another connection to a different db so that one can query multiple dbs. This was causing subtle issues with transactions.
Namely, if inside a transaction, a user would execute a write _without_ prefixing the tables with the custom db name (using one connection), and that the QE would also execute a read, this time _with_ the tables prefixed with the db_name (using another connection), this would cause two different connections to do READ/WRITES inside a transaction, causing locks as explained in this doc https://sqlite.org/isolation.html