Better typing on OUTPUT for SQL Server
When using the `OUTPUT` on SQL Server `INSERT`, with triggers, the trick
with return list should be typed correctly. Originally we just assumed
the types close enough, but of course miscalculated certain exceptions.
E.g. `decimal` was always `decimal(32,16)`, which would not fit
`decimal(26,0)`...

Here we extend the type hints to the column with optional parameters.