Optionally expose postgres driver
This lets us drop down to a lower level API when we need to, at little
maintenance cost. The main use case for this in the Prisma Migration
Engine is to access the full error objects.

I want to extend this to other connectors if this pans out as expected
on postgres.