sslaccept and sslmode option implies that SSL needs to be enabled
The sslmode itself on its own does not enable / enforce SSL to be used.
Today this means that in cases where no custom CA or client side certificate
is provided, no SSL is enabled.

In turn, this means that for servers that provide a valid TLS
certificate signed by a default system root need a custom CA to be
passed in, normally a path to the system CA store.

This however turns into a lot of custom configuration, since the CA
store location is highly operating system dependent.

One example of a provider using system root is PlanetScale, but also other cloud
databases like Amazon Aurora or Azure MySQL can be provisioned with a
valid system signed root certificate and would benefit from this
improvement.