More secure SQL Server default TLS setting
This means without defining the `encrypt` parameter, the current
behavior would mean we enable TLS for the login, and turn it off for all
other communication with the database.

The new default just keeps TLS on for the whole session. It's 2021, all
sites use HTTPS already and why would the database connection be different...